requirements.txt
asgiref==3.8.1
boto3==1.36.26
botocore==1.36.26
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
crispy-tailwind==1.0.3
cryptography==44.0.2
Django==5.1.3
django-allauth==65.4.1
django-appconf==1.0.6
django-compressor==4.5.1
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-jazzmin==3.0.1
django-storages==1.14.5
django-taggit==6.1.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
psycopg==3.2.5
psycopg-binary==3.2.5
psycopg-pool==3.2.5
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
razorpay==1.4.2
rcssmin==1.1.2
requests==2.32.3
requests-oauthlib==2.0.0
rjsmin==1.2.2
s3transfer==0.11.2
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.8.2
